style: make `inv lint` pass over the full CI scope
An earlier pass only checked `visionsim tasks.py`, but tasks.py's PYTHON_DIRS
also covers tests/, scripts/, docs/ and examples/ -- so `inv lint` was still
failing. Measured against the baseline: upstream/main has 1 error under the
same command, this branch had 89. All 88 were ours.

  - 29 PLW1510: subprocess.run without an explicit `check`. These tests
    deliberately do not raise on a non-zero exit -- they capture output and
    assert on a sentinel in stdout, which gives a far better failure message
    than CalledProcessError. `check=False` states that instead of leaving it
    implicit. Applied via an AST pass, not a regex, after a regex attempt
    produced `,,` on calls that already had a trailing comma.
  - 33 typing modernisations in tests/ and scripts/, matching the convention
    the package already uses everywhere else.
  - B017 in test_presets_are_immutable now asserts FrozenInstanceError rather
    than bare Exception. The test is KEPT: ThermalPreset instances are shared
    by resolve_material/resolve_vertex_materials, so frozen=True is what stops
    one caller poisoning the global library.
  - PLC0206 in thermal_assign: iterate .items() rather than re-indexing.
  - RUF059 unused unpacked values underscore-prefixed to state the intent.

+++ tests/test_heatsim_atlas.py
@@ -487,7 +487,7 @@ def test_rasterize_performance_sanity():
 
     # Build a grid of small quads (2 triangles each) tiling the UV unit square, so triangles are
     # spatially local (like real UV islands) rather than degenerate/overlapping garbage.
-    grid_n = int(math.ceil(math.sqrt(n_tris / 2)))
+    grid_n = math.ceil(math.sqrt(n_tris / 2))
     step = 1.0 / grid_n
     verts = []
     faces = []
